requirements.txt
couchbase
python-dotenv
pymongo
certifi